refactor: change extend function postion
