+++ src/base.mjs
@@ -1,5 +1,4 @@
-import Stanz from "./main.mjs";
-import { isxdata } from "./public.mjs";
+import Stanz, { isxdata } from "./main.mjs";
 
 const stanz = (data) => {
   return new Stanz(data);
