+++ src/public.mjs
@@ -1,7 +1,4 @@
-import Stanz from "./main.mjs";
 export const getRandomId = () => Math.random().toString(32).slice(2);
-
-export const isxdata = (val) => val instanceof Stanz;
 
 const objectToString = Object.prototype.toString;
 export const getType = (value) =>
